+++ src/libft/Makefile
@@ -24,8 +24,13 @@ SRCS = ft_memset.c\
 		ft_toupper.c\
 		ft_tolower.c\
 		ft_calloc.c\
+		ft_calloc2.c\
+		ft_calloc3.c\
+		ft_realloc2.c\
+		ft_realloc3.c\
 		ft_strdup.c\
 		ft_substr.c\
+		ft_charjoin.c\
 		ft_strjoin.c\
 		ft_strtrim.c\
 		ft_putchar_fd.c\
