+++ src/libft/ft_charjoin.c
@@ -1,0 +1,19 @@
+#include "../../includes/libft.h"
+
+char *ft_charjoin(char *old, char c)
+{
+	char *new;
+	int len;
+	int i;
+
+	len = ft_strlen(old);
+	new = (char *)malloc(sizeof(char) * (len + 2));
+	i = -1;
+	while (++i < len)
+		new[i] = old[i];
+	new[i++] = c;
+	new[i] = '\0';
+	free(old);
+	return (new);
+}
+
